grim/cleans up assistant_agent structure
