+++ langgraph.json
@@ -3,7 +3,7 @@
   "dockerfile_lines": [],
   "dependencies": ["."],
   "graphs": {
-    "agent": "./src/react_agent/graph.ts:graph"
+    "assistant_agent": "./src/assistant_agent/graph.ts:graph"
   },
   "env": ".env"
 }
