+++ package.json
@@ -23,7 +23,8 @@
     "@langchain/core": "^0.3.57",
     "@langchain/langgraph": "^0.3.0",
     "@langchain/mcp-adapters": "^0.6.0",
-    "langchain": "^0.3.27"
+    "langchain": "^0.3.27",
+    "zod": "^3.25.76"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3.1.0",
