+++ src/assistant_agent/configuration.ts
@@ -1,26 +1,14 @@
 /**
  * Define the configurable parameters for the agent.
  */
-import { Annotation } from "@langchain/langgraph";
 import { SYSTEM_PROMPT_TEMPLATE } from "./prompts.js";
 import { RunnableConfig } from "@langchain/core/runnables";
 import { DEFAULT_MODEL } from "../utils/constants.js";
-
-export const ConfigurationSchema = Annotation.Root({
-  /**
-   * The system prompt to be used by the agent.
-   */
-  systemPromptTemplate: Annotation<string>,
-
-  /**
-   * The name of the language model to be used by the agent.
-   */
-  model: Annotation<string>,
-});
+import { AssistantConfiguration } from "./schemas/configuration-schema.js";
 
 export function ensureConfiguration(
   config: RunnableConfig,
-): typeof ConfigurationSchema.State {
+): AssistantConfiguration {
   /**
    * Ensure the defaults are populated.
    */
