+++ src/assistant_agent/schemas/configuration-schema.ts
@@ -1,0 +1,8 @@
+import { z } from "zod";
+
+export const AssistantConfigurationSchema = z.object({
+  model: z.string().optional(),
+  systemPromptTemplate: z.string().optional(),
+});
+
+export type AssistantConfiguration = z.infer<typeof AssistantConfigurationSchema>;
