+++ src/assistant_agent/tools.ts
