+++ src/assistant_agent/utils.ts
